logic separation
