+++ anton_potapov/graph.hpp
@@ -11,6 +11,8 @@
 #include <utility>
 #include <vector>
 
+#include "graph_traverser.hpp"
+
 enum class EdgeColor { Gray, Green, Blue, Yellow, Red };
 
 std::string get_edge_color_string(const EdgeColor& color) {
@@ -104,28 +106,29 @@ class Graph {
 
   Graph& operator=(const Graph&) = delete;
 
-  Graph& operator=(Graph&& other_graph) {
-    other_graph.update_vertices_depth();
-    vertices_ = std::move(other_graph.vertices_);
-    edges_ = std::move(other_graph.edges_);
-    next_vertex_id_ = std::move(other_graph.next_vertex_id_);
-    next_edge_id_ = std::move(other_graph.next_edge_id_);
-    vertices_at_depth_ = std::move(other_graph.vertices_at_depth_);
-    return *this;
-  }
+  Graph& operator=(Graph&& other_graph);
 
   Graph(const Graph&) = delete;
 
-  Graph(Graph&& other_graph) {
-    other_graph.update_vertices_depth();
-    vertices_ = std::move(other_graph.vertices_);
-    edges_ = std::move(other_graph.edges_);
-    next_vertex_id_ = std::move(other_graph.next_vertex_id_);
-    next_edge_id_ = std::move(other_graph.next_edge_id_);
-    vertices_at_depth_ = std::move(other_graph.vertices_at_depth_);
-  }
+  Graph(Graph&& other_graph);
 
   ~Graph() = default;
+
+  const Vertex& get_vertex(const VertexId& id) const {
+    return vertices_.at(id);
+  }
+
+  Vertex& get_vertex(const VertexId& id) {
+    const auto& const_this = *this;
+    return const_cast<Vertex&>(const_this.get_vertex(id));
+  }
+
+  const Edge& get_edge(const EdgeId& id) const { return edges_.at(id); }
+
+  Edge& get_edge(const EdgeId& id) {
+    const auto& const_this = *this;
+    return const_cast<Edge&>(const_this.get_edge(id));
+  }
 
   int max_depth() {
     update_vertices_depth();
@@ -139,77 +142,20 @@ class Graph {
 
   const std::map<VertexId, Vertex>& vertices() const { return vertices_; }
 
-  const std::set<VertexId>& get_vertices_at_depth(int depth) {
-    update_vertices_depth();
-    const auto& const_this = *this;
-    return const_this.get_vertices_at_depth(depth);
-  }
+  const std::set<VertexId>& get_vertices_at_depth(int depth);
 
-  const std::set<VertexId>& get_vertices_at_depth(int depth) const {
-    return vertices_at_depth_.at(depth);
-  }
+  const std::set<VertexId>& get_vertices_at_depth(int depth) const;
 
-  bool is_vertex_exists(const VertexId& vertex_id) const {
-    return vertices_.find(vertex_id) != vertices_.end();
-  }
+  bool is_vertex_exists(const VertexId& vertex_id) const;
 
   bool is_connected(const VertexId& vertex1_id,
-                    const VertexId& vertex2_id) const {
-    assert(is_vertex_exists(vertex1_id) && "Vertex 1 doesn't exist");
-    assert(is_vertex_exists(vertex2_id) && "Vertex 2 doesn't exist");
-    if (vertex1_id == vertex2_id) {
-      for (const auto& vertex_edge_id :
-           get_vertex(vertex1_id).connected_edges()) {
-        const auto& vertex_edge = get_edge(vertex_edge_id);
-        if (vertex_edge.vertex1_id == vertex_edge.vertex2_id) {
-          return true;
-        }
-      }
-      return false;
-    } else {
-      for (const auto& vertex1_edge_id :
-           get_vertex(vertex1_id).connected_edges()) {
-        if (vertices_.find(vertex2_id)->second.has_edge_id(vertex1_edge_id)) {
-          return true;
-        }
-      }
-      return false;
-    }
-  }
+                    const VertexId& vertex2_id) const;
 
-  VertexId add_vertex() {
-    const VertexId new_vertex_id = get_next_vertex_id();
-    vertices_.emplace(new_vertex_id, new_vertex_id);
-    return new_vertex_id;
-  }
+  VertexId add_vertex();
 
   EdgeId add_edge(const VertexId& vertex1_id,
                   const VertexId& vertex2_id,
-                  const EdgeColor& edge_color = EdgeColor::Gray) {
-    assert(is_vertex_exists(vertex1_id) && "Vertex 1 doesn't exist");
-    assert(is_vertex_exists(vertex2_id) && "Vertex 2 doesn't exist");
-    assert(!is_connected(vertex1_id, vertex2_id) &&
-           "Vertices already connected");
-    assert(new_edge_color_is_correct(vertex1_id, vertex2_id, edge_color) &&
-           "the new edge's color is incorrect");
-
-    const EdgeId new_edge_id = get_next_edge_id();
-    edges_.emplace(new_edge_id,
-                   Edge(new_edge_id, vertex1_id, vertex2_id, edge_color));
-
-    get_vertex(vertex1_id).add_edge(new_edge_id);
-    if (vertex1_id != vertex2_id) {
-      get_vertex(vertex2_id).add_edge(new_edge_id);
-    }
-
-    is_depth_dirty_ = true;
-    if (updated_depth_ != INIT_DEPTH) {
-      updated_depth_ =
-          std::min(updated_depth_, get_new_depth(vertex1_id, vertex2_id));
-    }
-
-    return new_edge_id;
-  }
+                  const EdgeColor& edge_color = EdgeColor::Gray);
 
   std::string get_json_string() {
     update_vertices_depth();
@@ -238,53 +184,16 @@ class Graph {
     return json_stringstream.str();
   }
 
+  const VertexId& get_root_vertex_id() const {
+    return vertices_.begin()->first;
+  }
+
   void update_vertices_depth() {
     if (!is_depth_dirty_) {
       return;
     }
-    std::map<VertexId, int> depths;
-    std::queue<VertexId> bfs_queue;
-    std::set<VertexId> used;
-    if (updated_depth_ == INIT_DEPTH) {
-      const VertexId& first_vertex_id = vertices_.begin()->first;
-      depths.emplace(first_vertex_id, 0);
-      bfs_queue.push(first_vertex_id);
-      used.insert(first_vertex_id);
-    } else {
-      int max_correct_depth = updated_depth_ - 1;
-      for (const auto& vertex_id : vertices_at_depth_.at(max_correct_depth)) {
-        depths.emplace(vertex_id, max_correct_depth);
-        bfs_queue.push(vertex_id);
-        used.insert(vertex_id);
-      }
-      if (max_correct_depth > 0) {
-        for (const auto& vertex_id :
-             vertices_at_depth_.at(max_correct_depth - 1)) {
-          used.insert(vertex_id);
-        }
-      }
-    }
-
-    while (!bfs_queue.empty()) {
-      const VertexId& current_vertex_id = bfs_queue.front();
-      bfs_queue.pop();
-      for (const auto& connected_edge_id :
-           get_vertex(current_vertex_id).connected_edges()) {
-        const Edge& connected_edge = get_edge(connected_edge_id);
-        const VertexId& vertex1_id = connected_edge.vertex1_id;
-        const VertexId& vertex2_id = connected_edge.vertex2_id;
-        const VertexId& connected_vertex_id =
-            (current_vertex_id == vertex1_id ? vertex2_id : vertex2_id);
-        if (used.find(connected_vertex_id) == used.end()) {
-          used.insert(connected_vertex_id);
-          depths[connected_vertex_id] = depths[current_vertex_id] + 1;
-          bfs_queue.push(connected_vertex_id);
-        }
-      }
-    }
-
+    const auto& depths = GraphTraverser::dynamic_bfs(*this, updated_depth_);
     update_vertices_at_depth_map(depths);
-
     is_depth_dirty_ = false;
   }
 
@@ -296,22 +205,6 @@ class Graph {
   std::map<VertexId, Vertex> vertices_;
   std::map<EdgeId, Edge> edges_;
   std::map<int, std::set<VertexId>> vertices_at_depth_;
-
-  const Vertex& get_vertex(const VertexId& id) const {
-    return vertices_.at(id);
-  }
-
-  Vertex& get_vertex(const VertexId& id) {
-    const auto& const_this = *this;
-    return const_cast<Vertex&>(const_this.get_vertex(id));
-  }
-
-  const Edge& get_edge(const EdgeId& id) const { return edges_.at(id); }
-
-  Edge& get_edge(const EdgeId& id) {
-    const auto& const_this = *this;
-    return const_cast<Edge&>(const_this.get_edge(id));
-  }
 
   VertexId get_next_vertex_id() { return next_vertex_id_++; }
 
@@ -379,11 +272,4 @@ class Graph {
   }
 };
 
-VertexId get_random_vertex_id(const std::set<VertexId>& vertex_id_set) {
-  std::random_device rd;
-  std::mt19937 gen(rd());
-  std::uniform_int_distribution<> dist(0, (int)vertex_id_set.size() - 1);
-  auto vertex_id_set_it = vertex_id_set.begin();
-  std::advance(vertex_id_set_it, dist(gen));
-  return *vertex_id_set_it;
-}
+VertexId get_random_vertex_id(const std::set<VertexId>& vertex_id_set);
